Extract widget related functionality into DashboardWIdgetsResource.
